feat: separate global provider into separate context providers
