+++ app/components/Header/Menu/MobileMenu.tsx
@@ -2,7 +2,7 @@ import {Swiper, SwiperSlide} from 'swiper/react';
 import {A11y} from 'swiper/modules';
 
 import {Link, Drawer, Svg} from '~/components';
-import {useColorSwatches, useGlobal, useSettings} from '~/hooks';
+import {useColorSwatches, useMenu, useSettings} from '~/hooks';
 
 import {ProductItem} from '../../ProductItem';
 import type {UseMobileMenuReturn} from '../useMobileMenu';
@@ -24,7 +24,7 @@ export function MobileMenu({
   handleMobileSubmenu,
 }: MobileMenuProps) {
   const {header} = useSettings();
-  const {openSearch} = useGlobal();
+  const {openSearch} = useMenu();
   const swatchesMap = useColorSwatches();
 
   const {links: additionalLinks, menuItems, productsSlider} = {...header?.menu};
