+++ app/contexts/index.ts
@@ -1,2 +1,6 @@
+export * from './ContextsProvider';
 export * from './GlobalProvider';
 export * from './GroupingsProvider';
+export * from './MenuProvider';
+export * from './PromobarProvider';
+export * from './SettingsProvider';
