+++ app/hooks/index.ts
@@ -12,7 +12,9 @@ export * from './useGlobal';
 export * from './useIsHydrated';
 export * from './useLoadScript';
 export * from './useMarketingListSubscribe';
+export * from './useMenu';
 export * from './usePagination';
+export * from './usePreviewMode';
 export * from './usePromobar';
 export * from './useRootLoaderData';
 export * from './useSettings';
