+++ app/hooks/product/useAddToCart.ts
@@ -6,7 +6,7 @@ import type {
   SellingPlan,
 } from '@shopify/hydrogen/storefront-api-types';
 
-import {useGlobal, useSettings} from '~/hooks';
+import {useMenu, useSettings} from '~/hooks';
 
 /**
  * Add to cart hook
@@ -56,7 +56,7 @@ export function useAddToCart({
 }: UseAddToCartProps): UseAddToCartReturn {
   const {error, linesAdd, status} = useCart();
   const {product: productSettings} = useSettings();
-  const {openCart, openModal} = useGlobal();
+  const {openCart, openModal} = useMenu();
 
   const [isAdding, setIsAdding] = useState(false);
   const [isAdded, setIsAdded] = useState(false);
