+++ app/hooks/useCustomer.ts
@@ -2,7 +2,7 @@ import {useMemo} from 'react';
 import {useMatches} from '@remix-run/react';
 import type {Customer} from '@shopify/hydrogen/storefront-api-types';
 
-import {useGlobal} from '~/hooks';
+import {usePreviewMode} from '~/hooks';
 
 import type {RootLoaderData} from './useRootLoaderData';
 
@@ -17,7 +17,7 @@ import type {RootLoaderData} from './useRootLoaderData';
 
 export function useCustomer(): Customer | null | undefined {
   const [root, account] = useMatches();
-  const {previewModeCustomer} = useGlobal();
+  const {previewModeCustomer} = usePreviewMode();
   const customerFromRoot = (root?.data as RootLoaderData)?.customer;
   const customerFromAccount = (account?.data as {customer: Customer})?.customer;
   const isPreviewModeEnabled = (root?.data as RootLoaderData)
