+++ app/hooks/useGlobal.ts
@@ -1,15 +1,19 @@
+import {useMemo} from 'react';
+
 import type {GlobalContext} from '~/lib/types';
 import {useGlobalContext} from '~/contexts';
 
 /**
- * Hook for Global Provider, e.g. cartOpen, searchOpen, etc.
+ * Hook for Global Provider, e.g. isCartReady
  * @returns Global context
  */
 
 export function useGlobal(): GlobalContext['state'] & GlobalContext['actions'] {
-  const {state, actions} = useGlobalContext() as GlobalContext;
-  return {
-    ...state,
-    ...actions,
-  };
+  const {state, actions} = useGlobalContext();
+  return useMemo(() => {
+    return {
+      ...state,
+      ...actions,
+    };
+  }, [state, actions]);
 }
