+++ app/hooks/useSettings.ts
@@ -1,7 +1,8 @@
+import {useMemo} from 'react';
 import {useSiteSettings} from '@pack/react';
 
-import {useGlobal} from '~/hooks';
-import type {Settings, SiteSettings} from '~/lib/types';
+import {useSettingsContext} from '~/contexts';
+import type {Settings} from '~/lib/types';
 
 /**
  * Hook for getting settings from Pack's useSiteSettings hook, with a fallback to settings set in the global context provider to be used within error boundaries
@@ -15,7 +16,9 @@ import type {Settings, SiteSettings} from '~/lib/types';
  */
 
 export function useSettings(): Settings {
-  const siteSettings = useSiteSettings() as SiteSettings;
-  const {settings} = useGlobal();
-  return {...(siteSettings?.settings || settings)};
+  const siteSettings = useSiteSettings();
+  const {state} = useSettingsContext();
+  return useMemo(() => {
+    return {...(siteSettings?.settings || state.settings)};
+  }, [siteSettings?.settings, state.settings]);
 }
