+++ app/lib/customer/hooks/useCustomerDeleteAddress.ts
@@ -5,7 +5,7 @@ import {
   getCustomerAccessTokenFromLocalStorage,
   usePreviewModeCustomerFetch,
 } from '~/lib/customer';
-import {useGlobal} from '~/hooks';
+import {usePreviewMode} from '~/hooks';
 
 import {useFetcherStatus} from './useFetcherStatus';
 
@@ -15,7 +15,7 @@ interface FetcherData {
 }
 
 export function useCustomerDeleteAddress() {
-  const {isPreviewModeEnabled} = useGlobal();
+  const {isPreviewModeEnabled} = usePreviewMode();
   const fetcher = useFetcher({key: 'delete-address'});
   const {deletedCustomerAddressId, deleteErrors} = {
     ...(fetcher.data as FetcherData),
