+++ app/lib/customer/hooks/useCustomerOrders.ts
@@ -3,11 +3,11 @@ import {useFetcher, useLoaderData} from '@remix-run/react';
 import type {Order} from '@shopify/hydrogen-react/storefront-api-types';
 
 import {getCustomerAccessTokenFromLocalStorage} from '~/lib/customer';
-import {useGlobal} from '~/hooks';
+import {usePreviewMode} from '~/hooks';
 import type {loader} from '~/routes/($locale).account.orders._index';
 
 export function useCustomerOrders() {
-  const {isPreviewModeEnabled} = useGlobal();
+  const {isPreviewModeEnabled} = usePreviewMode();
   const {orders, errors} = useLoaderData<typeof loader>();
   const fetcher = useFetcher<{orders: Order[]}>({key: 'orders'});
