+++ app/lib/types/index.ts
@@ -1,5 +1,5 @@
+export type * from './context.types';
 export type * from './general.types';
-export type * from './global.types';
 export type * from './pack.types';
 export type * from './product.types';
 export type * from './store.types';
